Refactored column value stores to fix a few bugs and make the code more consistant.
